Fix typing for procedure and add charset options

+++ procedures/productsByInvoice/get-invoice-total.sql
@@ -4,7 +4,7 @@ DROP PROCEDURE IF EXISTS GetInvoiceTotal$$
 
 CREATE PROCEDURE GetInvoiceTotal(IN p_invoice_id INT)
 BEGIN
-    DECLARE total DECIMAL(10, 2) DEFAULT 0.00;
+    DECLARE total INT DEFAULT 0;
 
     SELECT SUM(P.price * PBI.quantity) INTO total
     FROM `products_by_invoice` PBI
